Revert some changes and fix a bug
